refactor(interpreter): split checkpoint and bytecode tests
