+++ src/interpreter/src/interpreter.rs
@@ -2044,532 +2044,10 @@ fn register_bytecode_function(
   Ok(output)
 }
 
-#[cfg(all(
-  test,
-  feature = "functions",
-  feature = "symbol_table",
-  feature = "f64"
-))]
-mod checkpoint_tests {
-  use super::*;
+#[cfg(test)]
+#[path = "interpreter/tests/mod.rs"]
+mod tests;
 
-  fn f64_value(value: &Ref<f64>) -> Value {
-    Value::F64(value.clone())
-  }
-
-  fn install_scalar(
-    interpreter: &Interpreter,
-    name: &str,
-    value: f64,
-  ) -> (ValRef, Ref<f64>) {
-    let backing = Ref::new(value);
-    let id = hash_str(name);
-    let symbols = interpreter.symbols();
-    let cell = symbols
-      .borrow_mut()
-      .insert(id, f64_value(&backing), true);
-    symbols
-      .borrow()
-      .dictionary
-      .borrow_mut()
-      .insert(id, name.to_string());
-    (cell, backing)
-  }
-
-  #[test]
-  fn interpreter_checkpoint_restores_private_state_and_recursive_child_identity() {
-    let mut root = Interpreter::new(1, 100);
-    root.ip = 7;
-    root.profile = true;
-    let (symbol_cell, symbol_backing) = install_scalar(&root, "kept", 1.0);
-    let symbol_cell_address = symbol_cell.addr();
-    let symbol_backing_address = symbol_backing.addr();
-    let symbol_backing_identity = ReactiveCellId::new(symbol_backing.id());
-    root.registers = vec![f64_value(&symbol_backing)];
-    root.constants = vec![f64_value(&Ref::new(2.0))];
-    root.out = f64_value(&symbol_backing);
-    root.code.push(MechSourceCode::String("before".to_string()));
-    root
-      .out_values
-      .borrow_mut()
-      .insert(hash_str("out"), f64_value(&symbol_backing));
-    *root.inline_eval_counter.borrow_mut() = 4;
-    *root.persistent_user_function_plan_depth.borrow_mut() = 2;
-    *root.deferred_expression_solve_depth.borrow_mut() = 3;
-    let context_binding_id = hash_str("checkpoint-context");
-    root.context_bindings.borrow_mut().insert(
-      context_binding_id,
-      RuntimeContextBinding {
-        name: "checkpoint-context".to_string(),
-        base_uri: "test://checkpoint".to_string(),
-      },
-    );
-    let original_manifests = root.module_manifests.borrow().clone();
-    #[cfg(feature = "trace")]
-    {
-      root.trace = true;
-      root.trace_to_stdout = false;
-    }
-    #[cfg(feature = "state_machines")]
-    {
-      let name = internal_pattern_value_identifier("checkpoint-fsm");
-      root.user_state_machines.borrow_mut().insert(
-        hash_str("checkpoint-fsm"),
-        FsmImplementation {
-          name: name.clone(),
-          input: Vec::new(),
-          start: Pattern::Wildcard,
-          arms: Vec::new(),
-        },
-      );
-      root.user_state_machine_specs.borrow_mut().insert(
-        hash_str("checkpoint-fsm"),
-        FsmSpecification {
-          name,
-          input: Vec::new(),
-          output: None,
-          states: Vec::new(),
-        },
-      );
-    }
-    #[cfg(feature = "invariant_define")]
-    let invariant_result = Ref::new(Value::Bool(Ref::new(true)));
-    #[cfg(feature = "invariant_define")]
-    let invariant_rhs = Ref::new(Value::F64(Ref::new(2.0)));
-    #[cfg(feature = "invariant_define")]
-    {
-      let invariant_id = hash_str("checkpoint-invariant");
-      let mut state = root.state.borrow_mut();
-      state.integrity_constraints.insert(
-        invariant_id,
-        IntegrityConstraint {
-          id: invariant_id,
-          name: "checkpoint invariant".to_string(),
-          expression: "kept <= 2.0".to_string(),
-          result: invariant_result.clone(),
-          lhs: Some(symbol_cell.clone()),
-          operator: Some(FormulaOperator::Comparison(
-            ComparisonOp::LessThanEqual,
-          )),
-          rhs: Some(invariant_rhs.clone()),
-          tokens: Vec::new(),
-        },
-      );
-    }
-
-    let state_address = root.state.addr();
-    let symbols_address = root.symbols().addr();
-    let symbol_dictionary_address = root.symbols().borrow().dictionary.addr();
-    let out_values_address = root.out_values.addr();
-    let inline_counter_address = root.inline_eval_counter.addr();
-    let context_bindings_address = root.context_bindings.addr();
-    let module_manifests_address = root.module_manifests.addr();
-    #[cfg(feature = "trace")]
-    let trace_events_address = root.trace_events.addr();
-    #[cfg(feature = "state_machines")]
-    let user_state_machines_address = root.user_state_machines.addr();
-    #[cfg(feature = "state_machines")]
-    let user_state_machine_specs_address = root.user_state_machine_specs.addr();
-    let sub_interpreters_address = root.sub_interpreters.addr();
-
-    let child_id = 2;
-    let grandchild_id = 3;
-    let mut child = Interpreter::new(child_id, 200);
-    child.ip = 8;
-    let (_child_cell, child_backing) = install_scalar(&child, "child", 20.0);
-    let mut grandchild = Interpreter::new(grandchild_id, 300);
-    grandchild.ip = 9;
-    let (_grandchild_cell, grandchild_backing) =
-      install_scalar(&grandchild, "grandchild", 30.0);
-    let grandchild_ref = Ref::new(Box::new(grandchild));
-    let grandchild_handle_address = grandchild_ref.addr();
-    child
-      .sub_interpreters
-      .borrow_mut()
-      .insert(grandchild_id, grandchild_ref);
-    let child_ref = Ref::new(Box::new(child));
-    let child_handle_address = child_ref.addr();
-    root
-      .sub_interpreters
-      .borrow_mut()
-      .insert(child_id, child_ref);
-
-    let checkpoint = root.checkpoint().unwrap();
-
-    root.id = 99;
-    root.ip = 99;
-    root.profile = false;
-    root.max_steps = 999;
-    root.registers.clear();
-    root.constants.clear();
-    root.code.push(MechSourceCode::String("after".to_string()));
-    root.out = Value::Empty;
-    root.state = Ref::new(ProgramState::new());
-    root.out_values = Ref::new(HashMap::new());
-    root.inline_eval_counter = Ref::new(99);
-    root.persistent_user_function_plan_depth = Ref::new(99);
-    root.deferred_expression_solve_depth = Ref::new(99);
-    root.context_bindings = Ref::new(HashMap::new());
-    root.module_manifests = Ref::new(ModuleManifestCatalog::new());
-    #[cfg(feature = "trace")]
-    {
-      root.trace = false;
-      root.trace_to_stdout = true;
-      root.trace_events = Ref::new(Vec::new());
-    }
-    #[cfg(feature = "state_machines")]
-    {
-      root.user_state_machines = Ref::new(HashMap::new());
-      root.user_state_machine_specs = Ref::new(HashMap::new());
-    }
-    *symbol_backing.borrow_mut() = 11.0;
-    *child_backing.borrow_mut() = 21.0;
-    *grandchild_backing.borrow_mut() = 31.0;
-    #[cfg(feature = "invariant_define")]
-    {
-      if let Value::Bool(value) = &*invariant_result.borrow() {
-        *value.borrow_mut() = false;
-      }
-      if let Value::F64(value) = &*invariant_rhs.borrow() {
-        *value.borrow_mut() = 99.0;
-      }
-    }
-
-    let removed_child = root
-      .sub_interpreters
-      .borrow_mut()
-      .remove(&child_id)
-      .unwrap();
-    {
-      let mut child = removed_child.borrow_mut();
-      child.id = 22;
-      child.ip = 88;
-      let removed_grandchild = child
-        .sub_interpreters
-        .borrow_mut()
-        .remove(&grandchild_id)
-        .unwrap();
-      drop(removed_grandchild);
-    }
-    drop(removed_child);
-    root
-      .sub_interpreters
-      .borrow_mut()
-      .insert(999, Ref::new(Box::new(Interpreter::new(999, 10))));
-
-    let held = symbol_backing.borrow();
-    let error = root.restore(checkpoint.clone()).unwrap_err();
-    assert_eq!(
-      error.kind_as::<ValueStateBorrowConflict>().unwrap().phase,
-      "restore-before"
-    );
-    assert_eq!(root.id, 99);
-    assert_eq!(root.ip, 99);
-    assert_ne!(root.state.addr(), state_address);
-    assert!(root.sub_interpreters.borrow().contains_key(&999));
-    assert!(!root.sub_interpreters.borrow().contains_key(&child_id));
-    assert_eq!(*held, 11.0);
-    drop(held);
-
-    root.restore(checkpoint).unwrap();
-
-    assert_eq!(root.id, 1);
-    assert_eq!(root.ip, 7);
-    assert!(root.profile);
-    assert_eq!(root.max_steps, 100);
-    assert_eq!(root.state.addr(), state_address);
-    assert_eq!(root.symbols().addr(), symbols_address);
-    assert_eq!(
-      root.symbols().borrow().dictionary.addr(),
-      symbol_dictionary_address
-    );
-    assert_eq!(root.out_values.addr(), out_values_address);
-    assert_eq!(root.inline_eval_counter.addr(), inline_counter_address);
-    assert_eq!(root.context_bindings.addr(), context_bindings_address);
-    assert_eq!(root.module_manifests.addr(), module_manifests_address);
-    assert_eq!(*root.module_manifests.borrow(), original_manifests);
-    assert_eq!(
-      root.context_bindings
-        .borrow()
-        .get(&context_binding_id)
-        .unwrap()
-        .base_uri,
-      "test://checkpoint",
-    );
-    #[cfg(feature = "trace")]
-    {
-      assert!(root.trace);
-      assert!(!root.trace_to_stdout);
-      assert_eq!(root.trace_events.addr(), trace_events_address);
-    }
-    #[cfg(feature = "state_machines")]
-    {
-      assert_eq!(root.user_state_machines.addr(), user_state_machines_address);
-      assert_eq!(
-        root.user_state_machine_specs.addr(),
-        user_state_machine_specs_address,
-      );
-      assert!(root
-        .user_state_machines
-        .borrow()
-        .contains_key(&hash_str("checkpoint-fsm")));
-      assert!(root
-        .user_state_machine_specs
-        .borrow()
-        .contains_key(&hash_str("checkpoint-fsm")));
-    }
-    #[cfg(feature = "invariant_define")]
-    {
-      let state = root.state.borrow();
-      let invariant_id = hash_str("checkpoint-invariant");
-      let constraint = state.integrity_constraints.get(&invariant_id).unwrap();
-      assert_eq!(constraint.id, invariant_id);
-      assert_eq!(constraint.name, "checkpoint invariant");
-      assert_eq!(constraint.expression, "kept <= 2.0");
-      assert_eq!(
-        constraint.operator,
-        Some(FormulaOperator::Comparison(
-          ComparisonOp::LessThanEqual,
-        )),
-      );
-      assert_eq!(constraint.result.addr(), invariant_result.addr());
-      assert_eq!(constraint.rhs.as_ref().unwrap().addr(), invariant_rhs.addr());
-      if let Value::Bool(value) = &*constraint.result.borrow() {
-        assert!(*value.borrow());
-      } else {
-        panic!("restored constraint result must remain bool");
-      }
-      if let Value::F64(value) = &*constraint.rhs.as_ref().unwrap().borrow() {
-        assert_eq!(*value.borrow(), 2.0);
-      } else {
-        panic!("restored constraint rhs must remain f64");
-      }
-    }
-    assert_eq!(root.sub_interpreters.addr(), sub_interpreters_address);
-    assert_eq!(root.registers.len(), 1);
-    assert_eq!(root.constants.len(), 1);
-    assert_eq!(root.code, vec![MechSourceCode::String("before".to_string())]);
-    assert_eq!(*root.inline_eval_counter.borrow(), 4);
-    assert_eq!(*root.persistent_user_function_plan_depth.borrow(), 2);
-    assert_eq!(*root.deferred_expression_solve_depth.borrow(), 3);
-    assert_eq!(symbol_cell.addr(), symbol_cell_address);
-    assert_eq!(symbol_backing.addr(), symbol_backing_address);
-    assert_eq!(
-      ReactiveCellId::new(symbol_backing.id()),
-      symbol_backing_identity
-    );
-    assert_eq!(*symbol_backing.borrow(), 1.0);
-    assert!(root.sub_interpreters.borrow().get(&999).is_none());
-
-    let restored_child = root
-      .sub_interpreters
-      .borrow()
-      .get(&child_id)
-      .cloned()
-      .unwrap();
-    assert_eq!(restored_child.addr(), child_handle_address);
-    let restored_grandchild = {
-      let child = restored_child.borrow();
-      assert_eq!(child.id, child_id);
-      assert_eq!(child.ip, 8);
-      assert_eq!(*child_backing.borrow(), 20.0);
-      child
-        .sub_interpreters
-        .borrow()
-        .get(&grandchild_id)
-        .cloned()
-        .unwrap()
-    };
-    assert_eq!(restored_grandchild.addr(), grandchild_handle_address);
-    let grandchild = restored_grandchild.borrow();
-    assert_eq!(grandchild.id, grandchild_id);
-    assert_eq!(grandchild.ip, 9);
-    assert_eq!(*grandchild_backing.borrow(), 30.0);
-  }
-
-  #[cfg(feature = "compiler")]
-  #[test]
-  fn interpreter_checkpoint_rejects_retained_compiler_context() {
-    let mut interpreter = Interpreter::new(41, 100);
-    interpreter.context = Some(CompileCtx::new());
-    let error = match interpreter.checkpoint() {
-      Ok(_) => panic!("retained compiler context checkpoint unexpectedly succeeded"),
-      Err(error) => error,
-    };
-    let unsupported = error
-      .kind_as::<InterpreterCheckpointUnsupportedCompilerContext>()
-      .unwrap();
-    assert_eq!(unsupported.interpreter_id, 41);
-  }
-}
-
-#[cfg(all(test, feature = "program", feature = "functions", feature = "symbol_table", feature = "f64"))]
-mod bytecode_dependency_tests {
-  use super::*;
-
-  struct BytecodeDependencyTestFunction {
-    output: Value,
-  }
-
-  impl MechFunctionImpl for BytecodeDependencyTestFunction {
-    fn solve(&self) {}
-
-    fn out(&self) -> Value {
-      self.output.clone()
-    }
-
-    fn to_string(&self) -> String {
-      "bytecode-dependency-test".to_string()
-    }
-
-    fn transaction_state_values(&self) -> MResult<Vec<Value>> {
-      Ok(self.reactive_output_values())
-    }
-  }
-
-  #[cfg(feature = "compiler")]
-  impl MechFunctionCompiler for BytecodeDependencyTestFunction {
-    fn compile(&self, _ctx: &mut CompileCtx) -> MResult<Register> {
-      Ok(0)
-    }
-  }
-
-  fn bytecode_dependency_test_factory(
-    args: FunctionArgs,
-  ) -> MResult<Box<dyn MechFunction>> {
-    let output = match args {
-      FunctionArgs::Nullary(output)
-      | FunctionArgs::Unary(output, _)
-      | FunctionArgs::Binary(output, _, _)
-      | FunctionArgs::Ternary(output, _, _, _)
-      | FunctionArgs::Quaternary(output, _, _, _, _)
-      | FunctionArgs::Variadic(output, _) => output,
-    };
-
-    Ok(Box::new(BytecodeDependencyTestFunction { output }))
-  }
-
-  fn scalar(value: f64) -> (Value, ReactiveCellId) {
-    let cell = Ref::new(value);
-    let id = ReactiveCellId::new(cell.id());
-    (Value::F64(cell), id)
-  }
-
-  #[test]
-  fn bytecode_nullary_registration_has_no_inputs() {
-    let state = ProgramState::new();
-    let (output, output_cell) = scalar(1.0);
-
-    let result = register_bytecode_function(
-      &state,
-      bytecode_dependency_test_factory,
-      FunctionArgs::Nullary(output.clone()),
-    )
-    .unwrap();
-
-    let plan = state.plan.borrow();
-    let node = plan.node(0).unwrap();
-    assert_eq!(plan.len(), 1);
-    assert!(node.inputs.is_empty());
-    assert!(plan.reactive_consumers.is_empty());
-    assert!(plan.sampled_consumers.is_empty());
-    assert!(node.outputs.contains(&output_cell));
-    assert_eq!(result.reactive_cell_ids(), output.reactive_cell_ids());
-  }
-
-  #[test]
-  fn bytecode_unary_registration_indexes_operand() {
-    let state = ProgramState::new();
-    let (output, output_cell) = scalar(1.0);
-    let (input, input_cell) = scalar(2.0);
-
-    register_bytecode_function(
-      &state,
-      bytecode_dependency_test_factory,
-      FunctionArgs::Unary(output, input),
-    )
-    .unwrap();
-
-    let plan = state.plan.borrow();
-    let node = plan.node(0).unwrap();
-    assert_eq!(node.inputs.len(), 1);
-    assert_eq!(node.inputs[0].cell, input_cell);
-    assert_eq!(node.inputs[0].kind, ReactiveDependencyKind::Reactive);
-    assert_eq!(plan.reactive_consumers_for(input_cell), &[0]);
-    assert!(!node.inputs.iter().any(|dependency| dependency.cell == output_cell));
-    assert!(node.outputs.contains(&output_cell));
-  }
-
-  #[test]
-  fn bytecode_binary_registration_preserves_operand_order() {
-    let state = ProgramState::new();
-    let (output, _) = scalar(1.0);
-    let (lhs, lhs_cell) = scalar(2.0);
-    let (rhs, rhs_cell) = scalar(3.0);
-
-    register_bytecode_function(
-      &state,
-      bytecode_dependency_test_factory,
-      FunctionArgs::Binary(output, lhs, rhs),
-    )
-    .unwrap();
-
-    let plan = state.plan.borrow();
-    let node = plan.node(0).unwrap();
-    assert_eq!(
-      node.inputs.iter().map(|dependency| dependency.cell).collect::<Vec<_>>(),
-      vec![lhs_cell, rhs_cell],
-    );
-    assert!(node.inputs.iter().all(|dependency| {
-      dependency.kind == ReactiveDependencyKind::Reactive
-    }));
-    assert_eq!(plan.reactive_consumers_for(lhs_cell), &[0]);
-    assert_eq!(plan.reactive_consumers_for(rhs_cell), &[0]);
-  }
-
-  #[test]
-  fn bytecode_variadic_registration_preserves_all_inputs() {
-    let state = ProgramState::new();
-    let (output, _) = scalar(1.0);
-    let (first, first_cell) = scalar(2.0);
-    let (second, second_cell) = scalar(3.0);
-    let (third, third_cell) = scalar(4.0);
-
-    register_bytecode_function(
-      &state,
-      bytecode_dependency_test_factory,
-      FunctionArgs::Variadic(output, vec![first, second, third]),
-    )
-    .unwrap();
-
-    let plan = state.plan.borrow();
-    let node = plan.node(0).unwrap();
-    assert_eq!(
-      node.inputs.iter().map(|dependency| dependency.cell).collect::<Vec<_>>(),
-      vec![first_cell, second_cell, third_cell],
-    );
-  }
-
-  #[test]
-  fn bytecode_registration_deduplicates_alias_operands() {
-    let state = ProgramState::new();
-    let (output, _) = scalar(1.0);
-    let (input, input_cell) = scalar(2.0);
-
-    register_bytecode_function(
-      &state,
-      bytecode_dependency_test_factory,
-      FunctionArgs::Binary(output, input.clone(), input),
-    )
-    .unwrap();
-
-    let plan = state.plan.borrow();
-    let node = plan.node(0).unwrap();
-    assert_eq!(node.inputs.len(), 1);
-    assert_eq!(node.inputs[0].cell, input_cell);
-    assert_eq!(plan.reactive_consumers_for(input_cell), &[0]);
-  }
-}
 
 // Interpreter Errors
 // ----------------------------------------------------------------------------
@@ -2730,453 +2208,8 @@ impl MechErrorKind for NoStepsInPlanError {
   }
 }
 
-#[cfg(all(test, feature = "program", feature = "compiler", feature = "functions", feature = "variables", feature = "variable_define", feature = "variable_assign", feature = "assign", feature = "f64", feature = "math"))]
-mod reactive_turn_interpreter_state_tests {
-  use super::*;
-  const SOURCE: &str = "input := 1.0\n~a := 0.0\n~b := 0.0\na = input\nmiddle := a + 1.0\nb = middle\noutput := b + 1.0";
-  fn interpreter() -> Interpreter { let mut i=Interpreter::new_with_full_stdlib(1); let t=mech_syntax::parser::parse(SOURCE).unwrap(); i.interpret(&t).unwrap(); i }
-  fn value(i:&Interpreter,n:&str)->f64 {let value=i.symbols().borrow().get(hash_str(n)).expect("symbol").borrow().clone();*value.as_f64().expect("f64").borrow()}
-  fn cell(i:&Interpreter,n:&str)->ReactiveCellId {let v=i.symbols().borrow().get(hash_str(n)).expect("symbol").borrow().reactive_root_cell_ids(); assert_eq!(v.len(),1,"root cell");v[0]}
-  fn register(i:&Interpreter,c:ReactiveCellId)->ReactiveNodeId {let p=i.plan();let v=p.borrow().nodes.iter().filter(|n|n.kind==ReactiveNodeKind::Register&&n.outputs.contains(&c)).map(|n|n.id).collect::<Vec<_>>();assert_eq!(v.len(),1,"register");v[0]}
-  fn first(i:&mut Interpreter)->(ReactiveNodeId,ReactiveNodeId){assert_eq!((value(i,"input"),value(i,"a"),value(i,"middle"),value(i,"b"),value(i,"output")),(1.,1.,2.,2.,3.));let(input,a,b)=(cell(i,"input"),register(i,cell(i,"a")),register(i,cell(i,"b")));let input_value=i.symbols().borrow().get(hash_str("input")).unwrap().borrow().clone();*input_value.as_f64().unwrap().borrow_mut()=10.;let o=i.advance_reactive_turn(&[input]).unwrap();assert_eq!(o.register_commit.committed_nodes,vec![a]);assert_eq!(o.after_commit.pending_register_nodes,vec![b]);(a,b)}
-  #[test] fn reactive_turn_interpreter_state_persists_between_calls(){let mut i=interpreter();let(a,b)=first(&mut i);assert_eq!((value(&i,"a"),value(&i,"middle"),value(&i,"b"),value(&i,"output")),(10.,11.,2.,3.));assert!(i.has_pending_reactive_registers());let o=i.advance_reactive_turn(&[]).unwrap();assert_eq!(o.register_commit.committed_nodes,vec![b]);assert!(!o.register_commit.committed_nodes.contains(&a));assert_eq!((value(&i,"a"),value(&i,"middle"),value(&i,"b"),value(&i,"output")),(10.,11.,11.,12.));assert!(o.after_commit.pending_register_nodes.is_empty());assert!(!i.has_pending_reactive_registers());}
-  #[test] fn reactive_turn_interpreter_state_clear_plan_resets_pending(){let mut i=interpreter();first(&mut i);assert!(i.has_pending_reactive_registers());assert!(i.plan_len()>0);i.clear_plan();assert_eq!(i.plan_len(),0);assert!(!i.has_pending_reactive_registers());}
-  #[test] fn reactive_turn_interpreter_state_clone_preserves_pending(){let mut i=interpreter();first(&mut i);let c=i.clone();assert!(i.has_pending_reactive_registers());assert!(c.has_pending_reactive_registers());assert_eq!(i.plan_len(),c.plan_len());}
-}
 
-#[cfg(all(test, feature = "functions"))]
-mod compact_reactive_turn_checkpoint_tests {
-  use super::*;
-  use std::{cell::RefCell, rc::Rc};
 
-  struct CompactTestFunction {
-    name: &'static str,
-    output: Ref<usize>,
-    captures: Rc<RefCell<usize>>,
-    solves: Rc<RefCell<usize>>,
-    fail_on_solve: Option<usize>,
-    leak_borrow_on_failure: bool,
-  }
-
-  impl CompactTestFunction {
-    fn new(
-      name: &'static str,
-      output: Ref<usize>,
-      captures: Rc<RefCell<usize>>,
-      solves: Rc<RefCell<usize>>,
-    ) -> Self {
-      Self {
-        name,
-        output,
-        captures,
-        solves,
-        fail_on_solve: None,
-        leak_borrow_on_failure: false,
-      }
-    }
-
-    fn execute(&self) -> MResult<()> {
-      let solve = {
-        let mut solves = self.solves.borrow_mut();
-        *solves += 1;
-        *solves
-      };
-      *self.output.borrow_mut() += 1;
-      if self.fail_on_solve == Some(solve) {
-        if self.leak_borrow_on_failure {
-          let borrow = self.output.borrow_mut();
-          std::mem::forget(borrow);
-        }
-        return Err(MechError::new(
-          GenericError { msg: format!("deliberate {} execution failure", self.name) },
-          None,
-        ));
-      }
-      Ok(())
-    }
-  }
-
-  impl MechFunctionImpl for CompactTestFunction {
-    fn solve(&self) {}
-
-    fn solve_result(&self) -> MResult<()> {
-      self.execute()
-    }
-
-    fn solve_reactive(&self) -> MResult<ReactiveSolveStatus> {
-      self.execute()?;
-      Ok(ReactiveSolveStatus::Changed)
-    }
-
-    fn out(&self) -> Value {
-      Value::Index(self.output.clone())
-    }
-
-    fn transaction_state_values(&self) -> MResult<Vec<Value>> {
-      *self.captures.borrow_mut() += 1;
-      Ok(vec![Value::Index(self.output.clone())])
-    }
-
-    fn to_string(&self) -> String {
-      self.name.into()
-    }
-  }
-
-  #[cfg(feature = "compiler")]
-  impl MechFunctionCompiler for CompactTestFunction {
-    fn compile(&self, _ctx: &mut CompileCtx) -> MResult<Register> {
-      Ok(0)
-    }
-  }
-
-  fn function(
-    name: &'static str,
-    output: Ref<usize>,
-  ) -> (CompactTestFunction, Rc<RefCell<usize>>, Rc<RefCell<usize>>) {
-    let captures = Rc::new(RefCell::new(0));
-    let solves = Rc::new(RefCell::new(0));
-    (
-      CompactTestFunction::new(name, output, captures.clone(), solves.clone()),
-      captures,
-      solves,
-    )
-  }
-
-  fn add_reactive(
-    interpreter: &Interpreter,
-    function: CompactTestFunction,
-    input: &Ref<usize>,
-  ) -> ReactiveNodeId {
-    interpreter.plan().0.borrow_mut()
-      .register(Box::new(function), &[Value::Index(input.clone())])
-      .unwrap()
-  }
-
-  #[test]
-  fn reactive_turn_checkpoint_contains_only_scheduler_metadata_and_plan_handles() {
-    let interpreter = Interpreter::new(7, 100);
-    let plan = interpreter.plan();
-    let (function, _, _) = function("node", Ref::new(0));
-    plan.add_function(Box::new(function));
-
-    let checkpoint = interpreter.reactive_turn_checkpoint().unwrap();
-
-    assert_eq!(checkpoint.interpreter_id, 7);
-    assert_eq!(checkpoint.plan_node_len, 1);
-    assert_eq!(checkpoint.activation_registration_depth, 0);
-    assert_eq!(checkpoint.plan.0.addr(), plan.0.addr());
-    assert_eq!(checkpoint.plan.1.addr(), plan.1.addr());
-  }
-
-  #[test]
-  fn reactive_turn_checkpoint_does_not_capture_function_values() {
-    let interpreter = Interpreter::new(8, 100);
-    let (function, captures, _) = function("node", Ref::new(0));
-    interpreter.plan().add_function(Box::new(function));
-
-    interpreter.reactive_turn_checkpoint().unwrap();
-
-    assert_eq!(*captures.borrow(), 0);
-  }
-
-  #[test]
-  fn reactive_turn_checkpoint_restores_pending_registers_exactly() {
-    let mut interpreter = Interpreter::new(9, 100);
-    interpreter.reactive_turn_state.pending_register_nodes = vec![3, 5, 3];
-    let checkpoint = match interpreter.reactive_turn_checkpoint() {
-      Ok(_) => panic!("invalid pending registers unexpectedly checkpointed"),
-      Err(error) => error,
-    };
-    assert_eq!(checkpoint.kind_name(), "InterpreterReactiveTurnInvariant");
-
-    let (function, _, _) = function("register-shaped", Ref::new(0));
-    let node = interpreter.plan().add_function(Box::new(function));
-    interpreter.plan().0.borrow_mut().nodes[node].kind = ReactiveNodeKind::Register;
-    interpreter.reactive_turn_state.pending_register_nodes = vec![node, node];
-    let checkpoint = interpreter.reactive_turn_checkpoint().unwrap();
-    interpreter.reactive_turn_state.pending_register_nodes.clear();
-    interpreter.preflight_restore_reactive_turn(&checkpoint).unwrap();
-    interpreter.apply_restore_reactive_turn(&checkpoint);
-    assert_eq!(interpreter.reactive_turn_state.pending_register_nodes, vec![node, node]);
-  }
-
-  #[test]
-  fn reactive_turn_failure_truncates_appended_trace_events() {
-    let mut interpreter = Interpreter::new(10, 100);
-    #[cfg(feature = "trace")]
-    {
-    interpreter.trace = true;
-    interpreter.trace_to_stdout = false;
-    }
-    let output = Ref::new(0);
-    let (mut function, _, _) = function("trace-fail", output);
-    function.fail_on_solve = Some(1);
-    interpreter.plan().add_function(Box::new(function));
-    #[cfg(feature = "trace")]
-    let original_len = interpreter.trace_events.borrow().len();
-
-    interpreter.step(0, 1).unwrap_err();
-
-    #[cfg(feature = "trace")]
-    assert_eq!(interpreter.trace_events.borrow().len(), original_len);
-  }
-
-  #[test]
-  fn reactive_turn_owner_mismatch_fails_before_mutation() {
-    let first = Interpreter::new(11, 100);
-    let second = Interpreter::new(11, 100);
-    let checkpoint = first.reactive_turn_checkpoint().unwrap();
-
-    let error = second.preflight_restore_reactive_turn(&checkpoint).unwrap_err();
-
-    assert_eq!(error.kind_name(), "InterpreterReactiveTurnOwnerMismatch");
-  }
-
-  #[test]
-  fn reactive_turn_interpreter_id_mismatch_fails_before_mutation() {
-    let mut interpreter = Interpreter::new(12, 100);
-    let checkpoint = interpreter.reactive_turn_checkpoint().unwrap();
-    interpreter.id = 13;
-
-    let error = interpreter.preflight_restore_reactive_turn(&checkpoint).unwrap_err();
-
-    assert_eq!(error.kind_name(), "InterpreterReactiveTurnInvariant");
-    assert!(error.kind_message().contains("interpreter ID"));
-  }
-
-  #[test]
-  fn reactive_turn_plan_handle_mismatch_fails_before_mutation() {
-    let mut interpreter = Interpreter::new(14, 100);
-    let checkpoint = interpreter.reactive_turn_checkpoint().unwrap();
-    interpreter.state.borrow_mut().plan = Plan::new();
-
-    let error = interpreter.preflight_restore_reactive_turn(&checkpoint).unwrap_err();
-
-    assert_eq!(error.kind_name(), "InterpreterReactiveTurnInvariant");
-    assert!(error.kind_message().contains("plan handles"));
-  }
-
-  #[test]
-  fn reactive_turn_plan_length_mismatch_fails_before_mutation() {
-    let interpreter = Interpreter::new(15, 100);
-    let checkpoint = interpreter.reactive_turn_checkpoint().unwrap();
-    let (function, _, _) = function("tail", Ref::new(0));
-    interpreter.plan().add_function(Box::new(function));
-
-    let error = interpreter.preflight_restore_reactive_turn(&checkpoint).unwrap_err();
-
-    assert_eq!(error.kind_name(), "InterpreterReactiveTurnInvariant");
-    assert!(error.kind_message().contains("plan length"));
-  }
-
-  #[test]
-  fn reactive_turn_activation_depth_mismatch_fails_before_mutation() {
-    let interpreter = Interpreter::new(16, 100);
-    let checkpoint = interpreter.reactive_turn_checkpoint().unwrap();
-    interpreter.plan().push_activation_registration_scope(Vec::new());
-
-    let error = interpreter.preflight_restore_reactive_turn(&checkpoint).unwrap_err();
-
-    assert_eq!(error.kind_name(), "InterpreterReactiveTurnInvariant");
-    assert!(error.kind_message().contains("activation registration depth"));
-  }
-
-  #[test]
-  fn reactive_turn_saved_missing_pending_node_is_rejected() {
-    let interpreter = Interpreter::new(17, 100);
-    let mut checkpoint = interpreter.reactive_turn_checkpoint().unwrap();
-    checkpoint.reactive_turn_state.pending_register_nodes = vec![99];
-
-    let error = interpreter.preflight_restore_reactive_turn(&checkpoint).unwrap_err();
-
-    assert_eq!(error.kind_name(), "InterpreterReactiveTurnInvariant");
-    assert!(error.kind_message().contains("does not exist"));
-  }
-
-  #[test]
-  fn reactive_turn_saved_non_register_pending_node_is_rejected() {
-    let interpreter = Interpreter::new(18, 100);
-    let (function, _, _) = function("comb", Ref::new(0));
-    let node = interpreter.plan().add_function(Box::new(function));
-    let mut checkpoint = interpreter.reactive_turn_checkpoint().unwrap();
-    checkpoint.reactive_turn_state.pending_register_nodes = vec![node];
-
-    let error = interpreter.preflight_restore_reactive_turn(&checkpoint).unwrap_err();
-
-    assert_eq!(error.kind_name(), "InterpreterReactiveTurnInvariant");
-    assert!(error.kind_message().contains("not a register"));
-  }
-
-  #[test]
-  fn participating_interpreter_turn_leaves_rollback_to_coordinator() {
-    let mut interpreter = Interpreter::new(19, 100);
-    let input = Ref::new(1usize);
-    let output = Ref::new(4usize);
-    let (mut function, _, _) = function("failure", output.clone());
-    function.fail_on_solve = Some(1);
-    add_reactive(&interpreter, function, &input);
-    with_reactive_journal_participant(|mut participant| {
-      let mut services = NoMechExecutionServices;
-      interpreter
-        .advance_reactive_turn_participating(
-          &Value::Index(input).reactive_root_cell_ids(),
-          &mut participant,
-          &mut services,
-        )
-        .unwrap_err();
-      assert_eq!(*output.borrow(), 5);
-      participant.preflight_restore_before()?;
-      participant.apply_restore_before();
-      Ok(())
-    })
-    .unwrap();
-    assert_eq!(*output.borrow(), 4);
-  }
-
-  #[test]
-  fn ordinary_interpreter_reactive_turn_rolls_back_on_failure() {
-    let mut interpreter = Interpreter::new(20, 100);
-    let input = Ref::new(1usize);
-    let output = Ref::new(4usize);
-    let (mut function, _, _) = function("failure", output.clone());
-    function.fail_on_solve = Some(1);
-    add_reactive(&interpreter, function, &input);
-
-    interpreter.advance_reactive_turn(
-      &Value::Index(input).reactive_root_cell_ids(),
-    ).unwrap_err();
-
-    assert_eq!(*output.borrow(), 4);
-  }
-
-  #[test]
-  fn ordinary_interpreter_step_rolls_back_earlier_function_mutation() {
-    let mut interpreter = Interpreter::new(21, 100);
-    let first_output = Ref::new(1usize);
-    let second_output = Ref::new(2usize);
-    let (first, _, _) = function("first", first_output.clone());
-    let (mut second, _, _) = function("second", second_output.clone());
-    second.fail_on_solve = Some(1);
-    interpreter.plan().add_function(Box::new(first));
-    interpreter.plan().add_function(Box::new(second));
-
-    interpreter.step(0, 1).unwrap_err();
-
-    assert_eq!((*first_output.borrow(), *second_output.borrow()), (1, 2));
-  }
-
-  #[test]
-  fn repeated_whole_plan_step_restores_before_first_repetition() {
-    let mut interpreter = Interpreter::new(22, 100);
-    let output = Ref::new(10usize);
-    let (mut function, _, _) = function("repeat", output.clone());
-    function.fail_on_solve = Some(3);
-    interpreter.plan().add_function(Box::new(function));
-
-    interpreter.step(0, 4).unwrap_err();
-
-    assert_eq!(*output.borrow(), 10);
-  }
-
-  #[test]
-  fn indexed_step_captures_only_selected_function() {
-    let mut interpreter = Interpreter::new(23, 100);
-    let (first, first_captures, _) = function("first", Ref::new(1));
-    let (second, second_captures, _) = function("second", Ref::new(2));
-    interpreter.plan().add_function(Box::new(first));
-    interpreter.plan().add_function(Box::new(second));
-    with_reactive_journal_participant(|mut participant| {
-      let mut services = NoMechExecutionServices;
-      interpreter.step_reactive_turn_participating(
-        2,
-        1,
-        &mut participant,
-        &mut services,
-      )?;
-      assert_eq!(
-        (*first_captures.borrow(), *second_captures.borrow()),
-        (0, 1),
-      );
-      participant.commit();
-      Ok(())
-    })
-    .unwrap();
-  }
-
-  #[test]
-  fn reactive_turn_rollback_failure_retains_original_error() {
-    let mut interpreter = Interpreter::new(24, 100);
-    let output = Ref::new(0usize);
-    let (mut function, _, _) = function("borrow-leak", output);
-    function.fail_on_solve = Some(1);
-    function.leak_borrow_on_failure = true;
-    interpreter.plan().add_function(Box::new(function));
-
-    let error = interpreter.step(0, 1).unwrap_err();
-
-    assert_eq!(
-      error.kind_name(),
-      "ReactiveJournalAutomaticRollbackFailed",
-    );
-    let rollback = error
-      .kind_as::<ReactiveJournalAutomaticRollbackFailed>()
-      .unwrap();
-    let original_error = rollback.original_error.as_ref().unwrap();
-    assert!(
-      original_error.contains("InterpreterReactiveTurnRollbackFailed"),
-    );
-    assert!(
-      original_error.contains("deliberate borrow-leak execution failure"),
-    );
-    assert!(original_error.contains("ValueStateBorrowConflict"));
-    assert!(rollback.rollback_error.contains("ValueStateBorrowConflict"));
-  }
-
-  #[test]
-  fn reactive_turn_paths_preserve_plan_identity_without_full_checkpoints() {
-    let mut interpreter = Interpreter::new(25, 100);
-    let plan = interpreter.plan();
-    let plan_address = plan.0.addr();
-    let (function, _, _) = function("success", Ref::new(0));
-    let node = plan.add_function(Box::new(function));
-
-    interpreter.step(0, 2).unwrap();
-
-    assert_eq!(interpreter.plan().0.addr(), plan_address);
-    assert_eq!(interpreter.plan().borrow().nodes[node].id, node);
-  }
-}
-
-#[cfg(all(test, feature = "program", feature = "compiler", feature = "functions", feature = "symbol_table", feature = "variable_define", feature = "f64"))]
-mod decoded_variable_definition_symbol_metadata_tests {
-  use super::*;
-
-  #[test]
-  fn decoded_variable_definition_symbol_metadata_round_trips() {
-    let tree = mech_syntax::parser::parse("input := 1.0\n~state := 2.0").unwrap();
-    let mut source = Interpreter::new_with_full_stdlib(1);
-    source.interpret(&tree).unwrap();
-    let bytes = source.compile().unwrap();
-    let parsed = ParsedProgram::from_bytes(&bytes).unwrap();
-    let input_id = hash_str("input");
-    let state_id = hash_str("state");
-    assert!(parsed.symbols.contains_key(&input_id));
-    assert!(parsed.symbols.contains_key(&state_id));
-    assert_eq!(parsed.dictionary.get(&input_id).unwrap(), "input");
-    assert_eq!(parsed.dictionary.get(&state_id).unwrap(), "state");
-    assert!(!parsed.mutable_symbols.contains(&input_id));
-    assert!(parsed.mutable_symbols.contains(&state_id));
-    let mut decoded = Interpreter::new_with_full_stdlib(2);
-    decoded.run_program(&parsed).unwrap();
-    for (name, expected) in [("input", 1.0), ("state", 2.0)] {
-      let value = decoded.symbols().borrow().get(hash_str(name)).unwrap().borrow().clone();
-      assert_eq!(*value.as_f64().unwrap().borrow(), expected);
-    }
-    let state = decoded.state.borrow();
-    assert!(state.get_mutable_symbol(input_id).is_none());
-    assert!(state.get_mutable_symbol(state_id).is_some());
-  }
-}
 pub struct InterpreterExecution<'a> {
   interpreter: &'a Interpreter,
   services: StdRefCell<&'a mut dyn MechExecutionServices>,
@@ -3256,57 +2289,5 @@ impl Deref for InterpreterExecution<'_> {
 
   fn deref(&self) -> &Self::Target {
     self.interpreter
-  }
-}
-
-#[cfg(test)]
-mod execution_services_borrow_tests {
-  use super::*;
-
-  #[test]
-  fn nested_execution_service_access_returns_a_structured_error() {
-    let interpreter = Interpreter::new(7001, 100);
-    let mut services = NoMechExecutionServices;
-    let execution = InterpreterExecution::new(
-      &interpreter,
-      &mut services,
-    );
-
-    let error = execution
-      .with_services(|_| {
-        execution.with_services(|_| Ok(()))
-      })
-      .unwrap_err();
-
-    assert_eq!(error.kind_name(), "ExecutionServicesBorrowConflict");
-    assert!(error.kind_message().contains("with_services"));
-  }
-
-  #[test]
-  fn reentrant_interpreter_execution_returns_an_error_without_panicking() {
-    let interpreter = Interpreter::new(7002, 100);
-    let nested = Interpreter::new(7003, 100);
-    let mut services = NoMechExecutionServices;
-    let execution = InterpreterExecution::new(
-      &interpreter,
-      &mut services,
-    );
-
-    let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(
-      || {
-        execution.with_services(|_| {
-          execution.with_interpreter(
-            &nested,
-            |_| Ok(()),
-          )
-        })
-      },
-    ));
-
-    let error = result
-      .expect("reentrant service access must not panic")
-      .unwrap_err();
-    assert_eq!(error.kind_name(), "ExecutionServicesBorrowConflict");
-    assert!(error.kind_message().contains("with_interpreter"));
   }
 }
