+++ src/interpreter/src/interpreter/tests/mod.rs
@@ -1,0 +1,4 @@
+mod bytecode;
+mod checkpoint;
+mod execution_services;
+mod reactive_turn;
